adding map search and refactoring
