+++ contextily/tile_providers.py
@@ -13,3 +13,8 @@ ST_TERRAIN_LINES = 'http://tile.stamen.com/terrain-lines/tileZ/tileX/tileY.png'
 ST_TERRAIN_BACKGROUND = 'http://tile.stamen.com/terrain-background/tileZ/tileX/tileY.png'
 
 ST_WATERCOLOR = 'http://tile.stamen.com/watercolor/tileZ/tileX/tileY.png'
+
+# OpenStreetMap as an alternative
+OSM_A = 'http://a.tile.openstreetmap.org/tileZ/tileX/tileY.png'
+OSM_B = 'http://b.tile.openstreetmap.org/tileZ/tileX/tileY.png'
+OSM_C = 'http://c.tile.openstreetmap.org/tileZ/tileX/tileY.png'
